add power on fixup.
git-svn-id: file:///svn/pmon-all/pmon-all@131 214b0138-1524-0410-9122-e5cb4b5bc56c

+++ pmon/cmds/mycmd.c
@@ -124,7 +124,7 @@ str2addr(const char *nptr,char **endptr,int base)
 unsigned long long result;
 if(syscall_addrtype%4==0)
 {
-result=strtoul(nptr,endptr,base);
+result=(long)strtoul(nptr,endptr,base);
 }
 else
 {
